+++ swarmlib/cuckoosearch/main.py
@@ -73,18 +73,6 @@ def configure_parser(sub_parsers):
         type=float,
         default=.1,
         help='Fraction of nests that will be randomly abandoned after each iteration (default 0.1)')
-    parser.add_argument(
-        '-i',
-        '--interval',
-        type=int,
-        default=1000,
-        help='Interval between two animation frames in ms (default 1000)')
-    parser.add_argument(
-        '-c',
-        '--continuous',
-        default=False,
-        action='store_true',
-        help='Enable the algorithm to run continuously (default off)')
 
     parser.add_argument(
         'nests',