+++ swarmlib/pso/main.py
@@ -79,18 +79,6 @@ def configure_parser(sub_parsers):
         type=float,
         default=2.,
         help='Maximum absolute velocity that is allowed for a particle (default 2.0)')
-    parser.add_argument(
-        '-i',
-        '--interval',
-        type=int,
-        default=1000,
-        help='Interval between two animation frames in ms (default 1000)')
-    parser.add_argument(
-        '-c',
-        '--continuous',
-        default=False,
-        action='store_true',
-        help='Enable the algorithm to run continuously (default off)')
 
     parser.add_argument(
         'particles',